Work
I have created a basis on what the application will follow, the rubbish disposable application. However, I need to make it so it will allow the user to input their data into the back-end and saves this, and also I will need to do the same for some sort of furniture relocation service. I also will need to brush on the CSS to make it looks better. I have created an email option so they can email extra details which I need to be aware of when they request services. I have created the forms which they can fill in